+++ sdk/src/main/java/com/hedera/hashgraph/sdk/Client.java
@@ -36,6 +36,8 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
     static final int DEFAULT_MAX_ATTEMPTS = 10;
     static final Duration DEFAULT_MAX_BACKOFF = Duration.ofSeconds(8L);
     static final Duration DEFAULT_MIN_BACKOFF = Duration.ofMillis(250L);
+    static final Duration DEFAULT_MAX_NODE_BACKOFF = Duration.ofHours(1L);
+    static final Duration DEFAULT_MIN_NODE_BACKOFF = Duration.ofSeconds(8L);
     static final Duration DEFAULT_CLOSE_TIMEOUT = Duration.ofSeconds(30L);
     static final Duration DEFAULT_REQUEST_TIMEOUT = Duration.ofMinutes(2L);
 
@@ -74,7 +76,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
         this.mirrorNetwork = mirrorNetwork;
     }
 
-    private static ExecutorService createExecutor() {
+    static ExecutorService createExecutor() {
         var threadFactory = new ThreadFactoryBuilder()
             .setNameFormat("hedera-sdk-%d")
             .setDaemon(true)
@@ -621,7 +623,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
      * @return
      */
     public synchronized Duration getNodeMinBackoff() {
-        return network.getMinBackoff();
+        return network.getMinNodeBackoff();
     }
 
     /**
@@ -631,7 +633,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
      * @return
      */
     public synchronized Client setNodeMinBackoff(Duration minBackoff) {
-        network.setMinBackoff(minBackoff);
+        network.setMinNodeBackoff(minBackoff);
         return this;
     }
 
@@ -641,7 +643,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
      * @return
      */
     public synchronized Duration getNodeMaxBackoff() {
-        return network.getMaxBackoff();
+        return network.getMaxNodeBackoff();
     }
 
     /**
@@ -651,7 +653,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
      * @return
      */
     public synchronized Client setNodeMaxBackoff(Duration maxBackoff) {
-        network.setMaxBackoff(maxBackoff);
+        network.setMaxNodeBackoff(maxBackoff);
         return this;
     }