+++ sdk/src/main/java/com/hedera/hashgraph/sdk/ContractId.java
@@ -1,16 +1,21 @@
 package com.hedera.hashgraph.sdk;
 
+import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.hedera.hashgraph.sdk.proto.ContractID;
+import org.bouncycastle.util.encoders.Hex;
 
 import javax.annotation.Nonnegative;
 import javax.annotation.Nullable;
+import java.util.Arrays;
 import java.util.Objects;
+import java.util.regex.Pattern;
 
 /**
  * The ID for a smart contract instance on Hedera.
  */
 public class ContractId extends Key implements Comparable<ContractId> {
+    static final Pattern EVM_ADDRESS_REGEX = Pattern.compile("(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.([a-fA-F0-9]{40}$)");
     /**
      * The shard number
      */
@@ -32,6 +37,9 @@ public class ContractId extends Key implements Comparable<ContractId> {
     @Nullable
     private final String checksum;
 
+    @Nullable
+    public final byte[] evmAddress;
+
     public ContractId(@Nonnegative long num) {
         this(0, 0, num);
     }
@@ -47,19 +55,54 @@ public class ContractId extends Key implements Comparable<ContractId> {
         this.realm = realm;
         this.num = num;
         this.checksum = checksum;
+        this.evmAddress = null;
+    }
+
+    ContractId(@Nonnegative long shard, @Nonnegative long realm, byte[] evmAddress) {
+        this.shard = shard;
+        this.realm = realm;
+        this.evmAddress = evmAddress;
+        this.num = 0;
+        this.checksum = null;
     }
 
     public static ContractId fromString(String id) {
-        return EntityIdHelper.fromString(id, ContractId::new);
+        var match = EVM_ADDRESS_REGEX.matcher(id);
+        if (match.find()) {
+            return new ContractId(
+                Long.parseLong(match.group(1)),
+                Long.parseLong(match.group(2)),
+                Hex.decode(match.group(3))
+            );
+        } else {
+            return EntityIdHelper.fromString(id, ContractId::new);
+        }
     }
 
+    @Deprecated
     public static ContractId fromSolidityAddress(String address) {
         return EntityIdHelper.fromSolidityAddress(address, ContractId::new);
     }
 
+    public static ContractId fromEvmAddress(@Nonnegative long shard, @Nonnegative long realm, String evmAddress) {
+        return new ContractId(
+            shard,
+            realm,
+            Hex.decode(evmAddress.startsWith("0x") ? evmAddress.substring(2) : evmAddress)
+        );
+    }
+
     static ContractId fromProtobuf(ContractID contractId) {
         Objects.requireNonNull(contractId);
-        return new ContractId(contractId.getShardNum(), contractId.getRealmNum(), contractId.getContractNum());
+        if (contractId.hasEvmAddress()) {
+            return new ContractId(
+                contractId.getShardNum(),
+                contractId.getRealmNum(),
+                contractId.getEvmAddress().toByteArray()
+            );
+        } else {
+            return new ContractId(contractId.getShardNum(), contractId.getRealmNum(), contractId.getContractNum());
+        }
     }
 
     public static ContractId fromBytes(byte[] bytes) throws InvalidProtocolBufferException {
@@ -67,15 +110,23 @@ public class ContractId extends Key implements Comparable<ContractId> {
     }
 
     public String toSolidityAddress() {
-        return EntityIdHelper.toSolidityAddress(shard, realm, num);
+        if (evmAddress != null) {
+            return Hex.toHexString(evmAddress);
+        } else {
+            return EntityIdHelper.toSolidityAddress(shard, realm, num);
+        }
     }
 
     ContractID toProtobuf() {
-        return ContractID.newBuilder()
+        var builder = ContractID.newBuilder()
             .setShardNum(shard)
-            .setRealmNum(realm)
-            .setContractNum(num)
-            .build();
+            .setRealmNum(realm);
+        if (evmAddress != null) {
+            builder.setEvmAddress(ByteString.copyFrom(evmAddress));
+        } else {
+            builder.setContractNum(num);
+        }
+        return builder.build();
     }
 
     /**
@@ -111,16 +162,24 @@ public class ContractId extends Key implements Comparable<ContractId> {
 
     @Override
     public String toString() {
-        return EntityIdHelper.toString(shard, realm, num);
+        if (evmAddress != null) {
+            return "" + shard + "." + realm + "." + Hex.toHexString(evmAddress);
+        } else {
+            return EntityIdHelper.toString(shard, realm, num);
+        }
     }
 
     public String toStringWithChecksum(Client client) {
-        return EntityIdHelper.toStringWithChecksum(shard, realm, num, client, checksum);
+        if (evmAddress != null) {
+            throw new IllegalStateException("toStringWithChecksum cannot be applied to ContractId with evmAddress");
+        } else {
+            return EntityIdHelper.toStringWithChecksum(shard, realm, num, client, checksum);
+        }
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(shard, realm, num);
+        return Objects.hash(shard, realm, num, Arrays.hashCode(evmAddress));
     }
 
     @Override
@@ -134,7 +193,18 @@ public class ContractId extends Key implements Comparable<ContractId> {
         }
 
         ContractId otherId = (ContractId) o;
-        return shard == otherId.shard && realm == otherId.realm && num == otherId.num;
+        return shard == otherId.shard && realm == otherId.realm && num == otherId.num && evmAddressMatches(otherId);
+    }
+
+    private boolean evmAddressMatches(ContractId otherId) {
+        if ((evmAddress == null) != (otherId.evmAddress == null)) {
+            return false;
+        }
+        if (evmAddress != null) {
+            return Arrays.equals(evmAddress, otherId.evmAddress);
+        }
+        // both are null
+        return true;
     }
 
     @Override
@@ -148,6 +218,22 @@ public class ContractId extends Key implements Comparable<ContractId> {
         if (realmComparison != 0) {
             return realmComparison;
         }
-        return Long.compare(num, o.num);
+        int numComparison = Long.compare(num, o.num);
+        if (numComparison != 0) {
+            return numComparison;
+        }
+        return evmAddressCompare(o);
+    }
+
+    private int evmAddressCompare(ContractId o) {
+        int nullCompare = (evmAddress == null ? 0 : 1) - (o.evmAddress == null ? 0 : 1);
+        if (nullCompare != 0) {
+            return nullCompare;
+        }
+        if (evmAddress != null) {
+            return Hex.toHexString(evmAddress).compareTo(Hex.toHexString(o.evmAddress));
+        }
+        // both are null
+        return 0;
     }
 }