+++ sdk/src/main/java/com/hedera/hashgraph/sdk/Network.java
@@ -1,22 +1,16 @@
 package com.hedera.hashgraph.sdk;
 
-import com.google.common.collect.HashBiMap;
 import com.google.common.io.ByteStreams;
 import com.google.common.io.Resources;
 import com.google.protobuf.ByteString;
-import java8.util.stream.Collectors;
-import java8.util.stream.StreamSupport;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeoutException;
 
@@ -297,7 +291,6 @@ class Network extends ManagedNetwork<Network, AccountId, Node> {
     @Override
     protected Node createNodeFromNetworkEntry(Map.Entry<String, AccountId> entry) {
         return new Node(entry.getValue(), entry.getKey(), executor)
-            .setMinBackoff(minBackoff)
             .setVerifyCertificates(verifyCertificates);
     }