+++ CHANGELOG.md
@@ -12,6 +12,77 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Checksums.  As a consequence, all previously generated checksums for `testnet` or `previewnet` will now be
    regarded as incorrect.  Please generate new checksums for testnet and previewnet where necessary.
 
+## v2.9.0
+
+### Added
+
+ * `owner` field to `*Allowance` classes.
+ * `Executable.[set|get]GrpcDeadline()`
+
+### Fixed
+
+ * `AccountAllowanceAdjustTransaction` now deserializes correctly with `Transaction.fromBytes()`
+
+## v2.9.0-beta.1
+
+### Added
+
+ * `owner` field to `*Allowance` classes.
+ * `Executable.[set|get]GrpcDeadline()`
+
+### Fixed
+
+ * `AccountAllowanceAdjustTransaction` now deserializes correctly with `Transaction.fromBytes()`
+
+## v2.8.0
+
+### Added
+
+ * CREATE2 Solidity addresses can now be represented by a `ContractId` with `evmAddress` set.
+ * `ContractId.fromEvmAddress()`
+ * `ContractFunctionResult.stateChanges`
+ * `ContractFunctionResult.evmAddress`
+ * `ContractStateChange`
+ * `StorageChange`
+ * New response codes.
+ * `ChunkedTransaction.[set|get]ChunkSize()`, and changed default chunk size for `FileAppendTransaction` to 2048.
+
+### Fixed
+
+ * `TransactionId.setRegenerateTransactionId()`
+ * `Transaction.execute(client, timeout)`
+
+### Deprecated
+
+* `ContractId.fromSolidityAddress()`, use `ContractId.fromEvmAddress()` instead.
+
+## v2.8.0-beta.1
+
+### Added
+
+ * CREATE2 Solidity addresses can now be represented by a `ContractId` with `evmAddress` set.
+ * `ContractId.fromEvmAddress()`
+ * `ContractFunctionResult.stateChanges`
+ * `ContractFunctionResult.evmAddress`
+ * `ContractStateChange`
+ * `StorageChange`
+ * New response codes.
+ * `ChunkedTransaction.[set|get]ChunkSize()`, and changed default chunk size for `FileAppendTransaction` to 2048.
+ * `AccountAllowance[Adjust|Approve]Transaction`
+ * `AccountInfo.[hbar|token|tokenNft]Allowances`
+ * `[Hbar|Token|TokenNft]Allowance`
+ * `[Hbar|Token|TokenNft]Allowance`
+ * `TransferTransaction.set[Hbar|Token|TokenNft]TransferApproval()`
+
+### Fixed
+
+ * `TransactionId.setRegenerateTransactionId()`
+ * `Transaction.execute(client, timeout)`
+
+### Deprecated
+
+* `ContractId.fromSolidityAddress()`, use `ContractId.fromEvmAddress()` instead.
+
 ## v2.7.0
 
 ### Added
@@ -24,7 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  * Support for regenerating transaction IDs on demand if a request
-   responsed with `TRANSACITON_EXPIRED`
+   responds with `TRANSACTION_EXPIRED`
 
 ## v2.6.0